requirements.txt
numpy>=1.19.0
opencv-python>=4.5.0
mediapipe>=0.8.9
torch>=1.9.0
torchvision>=0.10.0
paho-mqtt>=1.5.1
pygame>=2.0.1
Pillow>=8.0.0
depth-anything @ git+https://github.com/LiheYoung/Depth-Anything.git